model named 'ParticipantGroup' added for group events and some other minor changes
